+++ migrationsTimescale/20210130000004-change-confidence-on-detections.js
@@ -1,0 +1,23 @@
+'use strict'
+module.exports = {
+  up: (queryInterface, Sequelize) => {
+    return queryInterface.changeColumn(
+      'detections',
+      'confidence',
+      {
+        type: Sequelize.REAL,
+        allowNull: false
+      }
+    )
+  },
+  down: (queryInterface, Sequelize) => {
+    return queryInterface.changeColumn(
+      'detections',
+      'confidence',
+      {
+        type: Sequelize.FLOAT,
+        allowNull: false
+      }
+    )
+  }
+}